fix: move label back to the main PoS new invoice page
